fluid movement

+++ src/Entity.cpp
@@ -5,42 +5,42 @@
 #include "Entity.h"
 
 namespace model {
-    Entity::Entity(double length, double height, double x_position, double y_position) : length(length), height(height),
+    Entity::Entity(float length, float height, float x_position, float y_position) : length(length), height(height),
                                                                                          x_position(x_position),
                                                                                          y_position(y_position) {
         /// exceptions to be added!!
     }
 
 
-    const double Entity::get_length() const {
+    const float Entity::get_length() const {
         return length;
     }
 
-    const double Entity::get_height() const {
+    const float Entity::get_height() const {
         return height;
     }
 
-    double Entity::get_x_position() const {
+    float Entity::get_x_position() const {
         return x_position;
     }
 
-    double Entity::get_y_position() const {
+    float Entity::get_y_position() const {
         return y_position;
     }
 
-    double Entity::get_max_x_position() const {
+    float Entity::get_max_x_position() const {
         return max_x_position;
     }
 
-    double Entity::get_min_x_position() const {
+    float Entity::get_min_x_position() const {
         return min_x_position;
     }
 
-    double Entity::get_max_y_position() const {
+    float Entity::get_max_y_position() const {
         return max_y_position;
     }
 
-    double Entity::get_min_y_position() const {
+    float Entity::get_min_y_position() const {
         return min_y_position;
     }
