+++ src/StationaryEntity.cpp
@@ -6,8 +6,11 @@
 
 
 namespace model {
-    StationaryEntity::StationaryEntity(double length, double height, double x_position, double y_position) :
+    StationaryEntity::StationaryEntity(float length, float height, float x_position, float y_position) :
             Entity(length, height, x_position, y_position) {}
 
 }
+
+
+
