+++ src/Transformation.cpp
@@ -15,11 +15,11 @@ Transformation& Transformation::get_instance() {
 
 sf::Vector2f Transformation::co_to_pixels(std::shared_ptr<const model::Entity> entity, std::shared_ptr<const sf::RenderWindow> window) {
     /// make float for pixels??
-    double co_x_max = entity->get_max_x_position();
-    double co_y_max = entity->get_max_y_position();
+    float co_x_max = entity->get_max_x_position();
+    float co_y_max = entity->get_max_y_position();
 
-    double co_x = entity->get_x_position();
-    double co_y = entity->get_y_position();
+    float co_x = entity->get_x_position();
+    float co_y = entity->get_y_position();
 
     int pixel_x_max = window->getSize().x;
     int pixel_y_max = window->getSize().y;
@@ -27,8 +27,8 @@ sf::Vector2f Transformation::co_to_pixels(std::shared_ptr<const model::Entity> e
     float pixel_x;
     float pixel_y;
 
-    double entity_pixel_length = (entity->get_length() / co_x_max) * (pixel_x_max / 2);
-    double entity_pixel_height = round((entity->get_height() / co_y_max) * (pixel_y_max / 2) );
+    float entity_pixel_length = (entity->get_length() / co_x_max) * (pixel_x_max / 2);
+    float entity_pixel_height = round((entity->get_height() / co_y_max) * (pixel_y_max / 2) );
 
     if(co_x >= 0){
         pixel_x = (int) (round((co_x / co_x_max) * (pixel_x_max / 2) + (pixel_x_max / 2))) ;
